LFG-bot/requirements.txt
# LFG-bot Dependencies

# Exchange connectivity (using CCXT)
ccxt>=4.0.0

# Data handling
pandas>=2.0.0

# Technical analysis
ta>=0.11.0

# WebSocket for real-time data
websockets>=12.0

# Environment variables
python-dotenv>=1.0.0

# HTTP requests (backup)
requests>=2.31.0
